# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ESG ProofOps API
  version: 1.0.0
  description: Implementation contract, not a deployed service. Domain v2.0 is authoritative.
servers:
- url: /
paths:
...
  /v1/runs/{run_id}/claims:
    get:
      operationId: claims_list
      summary: claims_list
      x-minimum-role: viewer
      x-rate-limit: 120/min/user
      x-idempotency-required: false
      security:
      - sessionCookie: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimSummaryPage'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden or CSRF failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found or tenant mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict / gate blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Stale revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Schema or domain validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate/budget limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: cursor
        in: query
        schema:
          type: string
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 50
      - name: track
        in: query
        schema:
          type: string
          enum:
          - goal
          - performance
          - management
      - name: grade
        in: query
        schema:
          type: string
          enum:
          - E0
          - E1
          - E2
          - E3
      - name: review_status
        in: query
        schema:
          type: string
          enum:
          - auto_confirmed
          - needs_review
          - human_confirmed
  /v1/runs/{run_id}/claims/{claim_id}:
    get:
      operationId: claim_get
      summary: claim_get
      x-minimum-role: viewer
      x-rate-limit: 120/min/user
      x-idempotency-required: false
      security:
      - sessionCookie: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimDetail'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden or CSRF failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found or tenant mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict / gate blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Stale revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Schema or domain validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate/budget limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: claim_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-proofops_session
  schemas:
    Error:
      type: object
      additionalProperties: false
      properties:
        error:
          type: object
          additionalProperties: false
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
              format: uuid
            retryable: &id003
              type: boolean
            details:
              type: object
              additionalProperties: false
              properties:
                field:
                  type: string
                reason:
                  type: string
                current_revision:
                  type: integer
                  minimum: 1
                retry_after_seconds:
                  type: integer
                  minimum: 0
          required:
          - code
          - message
          - request_id
          - retryable
      required:
      - error
...
    Bbox:
      type: array
      items:
        type: number
      minItems: 4
      maxItems: 4
      description: Canonical top-left display points, x0,y0,x1,y1; missing geometry is null, not zero.
    SourceRef:
      type: object
      additionalProperties: false
      properties:
        source_id:
          type: string
          format: uuid
        document_version_id:
          type: string
          format: uuid
        parse_manifest_id:
          type: string
          format: uuid
        page_num:
          type: integer
          minimum: 1
        printed_page_label:
          anyOf:
          - type: string
          - type: 'null'
        bbox:
          anyOf:
          - $ref: '#/components/schemas/Bbox'
          - type: 'null'
        raw_text_sha256: *id002
        quote:
          type: string
        char_start:
          type: integer
          minimum: 0
        char_end:
          type: integer
          minimum: 0
        location_quality:
          type: string
          enum:
          - located
          - unlocated
          - unreadable
        verification_state:
          type: string
          enum:
          - verified
          - candidate
          - rejected
      required:
      - source_id
      - document_version_id
      - parse_manifest_id
      - page_num
      - printed_page_label
      - bbox
      - raw_text_sha256
      - quote
      - char_start
      - char_end
      - location_quality
      - verification_state
    Element:
      type: object
      additionalProperties: false
      properties:
        element_id:
          type: string
        state:
          type: string
          enum:
          - present
          - absent
          - unknown
          - conflict
          - not_applicable
        evidence_refs:
          type: array
          items:
            $ref: '#/components/schemas/SourceRef'
        normalized_value:
          anyOf:
          - type: string
          - type: 'null'
        credited_from:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
        reason_code:
          anyOf:
          - type: string
          - type: 'null'
      required:
      - element_id
      - state
      - evidence_refs
      - normalized_value
      - credited_from
      - reason_code
      allOf:
      - if:
          properties:
            state:
              const: present
        then:
          properties:
            evidence_refs:
              minItems: 1
    Decision:
      type: object
      additionalProperties: false
      properties:
        decision_revision:
          type: integer
          minimum: 1
        tag_revision:
          type: integer
          minimum: 1
        decision_status:
          type: string
          enum:
          - decided
          - blocked_evidence
          - blocked_rule_gap
          - not_applicable
          - not_run
        evidence_grade:
          anyOf:
          - type: string
            enum:
            - E0
            - E1
            - E2
            - E3
          - type: 'null'
        label:
          anyOf:
          - type: string
            enum:
            - SUBSTANTIATED
            - INCOMPLETE
            - UNSUBSTANTIATED
          - type: 'null'
        sublabel:
          anyOf:
          - type: string
            enum:
            - PERF
            - IMPL
          - type: 'null'
        review_status:
          type: string
          enum:
          - auto_confirmed
          - needs_review
          - human_confirmed
        missing_elements:
          type: array
          items:
            type: string
        rule_ids:
          type: array
          items:
            type: string
        rule_pack_sha256: *id002
        semantic_hash: *id002
        gap_ids:
          type: array
          items:
            type: string
      required:
      - decision_revision
      - tag_revision
      - decision_status
      - evidence_grade
      - label
      - sublabel
      - review_status
      - missing_elements
      - rule_ids
      - rule_pack_sha256
      - semantic_hash
      - gap_ids
      allOf:
      - if:
          properties:
            decision_status:
              const: decided
        then:
          properties:
            evidence_grade:
              type: string
              enum:
              - E0
              - E1
              - E2
              - E3
            label:
              type: string
              enum:
              - SUBSTANTIATED
              - INCOMPLETE
              - UNSUBSTANTIATED
        else:
          properties:
            evidence_grade:
              type: 'null'
            label:
              type: 'null'
            sublabel:
              type: 'null'
      - if:
          properties:
            evidence_grade:
              const: E0
        then:
          properties:
            label:
              const: UNSUBSTANTIATED
      - if:
          properties:
            evidence_grade:
              const: E1
        then:
          properties:
            label:
              const: INCOMPLETE
      - if:
          properties:
            evidence_grade:
              const: E2
        then:
          properties:
            label:
              const: INCOMPLETE
      - if:
          properties:
            evidence_grade:
              const: E3
        then:
          properties:
            label:
              const: SUBSTANTIATED
    ClaimSummary:
      type: object
      additionalProperties: false
      properties:
        claim_id:
          type: string
          format: uuid
        page_num:
          type: integer
          minimum: 1
        quote:
          type: string
        track:
          anyOf:
          - type: string
            enum:
            - goal
            - performance
            - management
          - type: 'null'
        topic_ids:
          type: array
          items:
            type: string
        decision:
          anyOf:
          - $ref: '#/components/schemas/Decision'
          - type: 'null'
        revision:
          type: integer
          minimum: 1
      required:
      - claim_id
      - page_num
      - quote
      - track
      - topic_ids
      - decision
      - revision
    AssuranceMatch:
      type: object
      additionalProperties: false
      properties:
        status:
          type: string
          enum:
          - covered
          - not_covered
          - undetermined
        level:
          anyOf:
          - type: string
            enum:
            - limited
            - reasonable
            - none
          - type: 'null'
        provider:
          anyOf:
          - type: string
          - type: 'null'
        statement_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
        metric_match:
          type: string
          enum:
          - 'yes'
          - 'no'
          - unknown
        period_match:
          type: string
          enum:
          - 'yes'
          - 'no'
          - unknown
        boundary_match:
          type: string
          enum:
          - 'yes'
          - 'no'
          - unknown
        evidence_refs:
          type: array
          items:
            $ref: '#/components/schemas/SourceRef'
      required:
      - status
      - level
      - provider
      - statement_id
      - metric_match
      - period_match
      - boundary_match
      - evidence_refs
    ClaimDetail:
      type: object
      additionalProperties: false
      properties:
        claim:
          $ref: '#/components/schemas/ClaimSummary'
        source_refs:
          type: array
          items:
            $ref: '#/components/schemas/SourceRef'
        elements:
          type: array
          items:
            $ref: '#/components/schemas/Element'
        assurance:
          $ref: '#/components/schemas/AssuranceMatch'
        replicate_request_ids:
          type: array
          items:
            type: string
        packet_sha256: *id002
        suggestion:
          anyOf:
          - type: string
          - type: 'null'
        basis_refs:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              standard:
                type: string
              clause:
                anyOf:
                - type: string
                - type: 'null'
              summary:
                type: string
              verification_status:
                type: string
                enum:
                - verified
                - unverified
                - unlicensed
            required:
            - standard
            - clause
            - summary
            - verification_status
      required:
      - claim
      - source_refs
      - elements
      - assurance
      - replicate_request_ids
      - packet_sha256
      - suggestion
      - basis_refs
...
    ClaimSummaryPage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ClaimSummary'
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
        snapshot_epoch:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
      required:
      - items
      - next_cursor
      - snapshot_epoch
...